requirements.txt
# Core dependencies for FE-EKG MVP
# Graph database clients (dual backend support)
agraph-python>=101.0.6  # AllegroGraph (RDF/SPARQL)
pycurl>=7.45.0          # Required by agraph-python for HTTPS
neo4j>=5.0.0            # Neo4j (Property Graph)
py2neo>=2021.2.3        # Neo4j OGM (alternative)

# RDF manipulation
rdflib>=6.0.0

# Data handling
pandas>=1.5.0
numpy>=1.23.0

# Graph visualization
networkx>=3.0
matplotlib>=3.5.0

# Environment variables
python-dotenv>=0.19.0

# Date/time utilities
python-dateutil>=2.8.0

# HTTP requests (for future news fetching)
requests>=2.28.0

# JSON handling
ujson>=5.0.0

# REST API
Flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0  # Production WSGI server

# LLM/Nemotron Integration (NVIDIA NIM)
# No additional packages needed - uses requests for API calls
# Optional: install openai package for alternative API clients
# openai>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
